* Ο κώδικας έγινε περισσότερο Java 8. Πολλές αλλαγές πραγματοποιήθηκαν και ελπίζω να μην οδηγήσουν σε bugs ;-) ~ Στον «Οδηγό Τιμολογίου» διορθώθηκε ένα bug το οποίο εμφάνιζε εσφαλμένα αποτελέσματα, αν επιλέγονταν μια μη επιτρεπόμενη επιλογή στα comboboxes. Π.χ. επιλέγονταν δαπάνη «Κατασκευή Έργου» ενώ προμηθευτής ήταν ο «Στρατός». * Αλλαγή εικονιδίου εφαρμογής ~ Διόρθωση στον installer, ώστε η συντόμευση να καλεί απευθείας το .jar αρχείο.

+++ src/cost/ContentItem.java
@@ -3,6 +3,6 @@ package cost;
 import common.*;
 
 public class ContentItem extends HashObject {
-	public ContentItem() { super.put("Πλήθος", 1); }
+	public ContentItem() { super.put("Πλήθος", (byte) 1); }
 	public String toString() { return (String) get("Δικαιολογητικό"); }
 }
